move all footer pages to routes dir
