+++ scruter-nextjs/app/(routes)/contributors/layout.tsx
