+++ scruter-nextjs/app/(routes)/contributors/page.tsx
@@ -2,7 +2,7 @@
 
 import React, { useEffect, useState } from 'react';
 import axios from 'axios';
-import '../globals.css'; // Ensure your global styles are imported
+import '../../globals.css'; // Ensure your global styles are imported
 import Navbar from '@/components/NavBars&Footers/navbar';
 import { Providers } from '@/lib/providers';
 import Footer from '@/components/NavBars&Footers/footer';
