+++ scruter-nextjs/app/(routes)/policy/page.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import '../globals.css'; // Ensure your global styles are imported
+import '../../globals.css'; // Ensure your global styles are imported
 import Navbar from '@/components/NavBars&Footers/navbar';
 import { Providers } from '@/lib/providers';
 import Footer from '@/components/NavBars&Footers/footer';
