+++ scruter-nextjs/app/(routes)/team/page.tsx
@@ -2,7 +2,7 @@
 
 import { useEffect } from 'react';
 import Image from 'next/image';
-import '../globals.css';
+import '../../globals.css';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import {
   faInstagram,
