+++ scruter-nextjs/app/(routes)/terms/layout.tsx
